Don't send back mode to confirm
Previously the client sent the mode received from the client back to confirm. I did this because I thought sending two messages in the same direction
simultaneously would combine them and require parsing between methods which would harm the clarity of the code. However I have since learned this is
not the case so removing the confirmation means we don't lose 2 round trips, the mode and username is just added to the start of the client message
but is parsed separately.